rename istio cert pkg
